Moved Minechaft on another repository
Minechaft game project become too big for stay in this repository, it's
better to make it's own repository for a better development, it will
also allow me to make different versions of this game without problems
of organization. A Hardcore mode will be planned in a future version of
the game. The game will be copied in another directory for do this.

Also thanks for anyone that beta-tested the game. It was hard work to do
and this game is one of my biggest projects with PC-9000